# facturation-planning-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour gérer la facturation et les plannings",
        "title": "Facturation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/salaries": {
            "get": {
                "description": "Retourne la liste des salariés enregistrés",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salariés"
                ],
                "summary": "Récupérer tous les salariés",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Salarie"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Ajoute un nouveau salarié à la base de données",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Salariés"
                ],
                "summary": "Créer un salarié",
                "parameters": [
                    {
                        "description": "Détails du salarié",
                        "name": "salarie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Salarie"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Salarie"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Salarie": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "jean.dupont@example.com"
                },
                "entreprise_id": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nom": {
                    "type": "string",
                    "example": "Jean Dupont"
                },
                "telephone": {
                    "type": "string",
                    "example": "+33123456789"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                }
            }
        }
    }
}